gcp/functions/blockables/requirements.txt
functions-framework==3.*
Flask
google-cloud-firestore
requests